[refactor] Refactor GetHTTPRules. (#252)
In writing an upcoming rule for AIP-156, I realized I have cases
where I *only* care about the URI, and not the method. Getting
this is a huge pain.

I have therefore refactored `GetHTTPRules` to return a custom
struct instead of the underlying proto annotation. The custom
struct has `Method` and `URI` fields that correspond to the
values on the annotation.

In the refactor to the standard method rules, the change clearly
does a better job of expressing intent: `if r.Method == "POST"`
is much clearer than `if r.GetPost() != ""`.