chore(deps): switch to new lro stub pkg (#1039)
Renovate tried to update `genproto` in #1037, but that change includes the deprecation and aliasing of the `longrunning` package to its new location in `cloud.google.com/go/longrunning/autogen/longrunningpb`. The deprecation notices on `genproto` triggered linter warnings. So rather than have the Go linter be upset about the deprecated deps on every PR, I am fixing it here with the genproto update, and subsequent direct dependency on the new lro stubs package.