fix: use openapi content schema, misc fixes
This contains a series of fixes to get aepcli to work
off of parsing the unmodified roblox api.

A few areas of note:

aepcli was using incorrect schemas previously, which assumed
the schema was directly under the response rather than under
a content map.

the help text was not very legible when called from the
api name, making that into a list makes it possible to find
the resources.

There are no ways to handle singletons nor custom methods 
currently - therefore those are skipped.

There are also scenarios where additional dereferencing
is needed - so added support for arbitrary dereferencing.

As a shortcut, fields that are references in a resource
schema are just defaulted to being handled as arbtirary json
objects. In the future we can add more nested validation.