feat(api): refactor to generate parents/patterns/children
The previous version of the API struct required that
pointers to parents and children were also passed - this
isn't strictly possible with value serialization from Json / Yaml,
resulting in significantly complexity around deduplication of objects.

In addition, there were default values in loaded resources missing (such as path).

To resolve this, many of the api.Resource fields were turned semi-private, generated
on load, loading from an API associated with the resources. By lazy-generating these
values, it enables a minimal set of values to be specified in markup formats, while
still enabling these structs to be used in the code base itself. Ultimately, this
prevents the need to define an on-disk resource format and an in-memory format.